Fix smaller bugs
