+++ .idea/markdown.xml
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project version="4">
+  <component name="MarkdownSettings">
+    <option name="hideErrorsInCodeBlocks" value="true" />
+  </component>
+</project>
