+++ CITATION.cff
@@ -6,4 +6,4 @@ authors:
     orcid: https://orcid.org/0000-0003-4562-5816
 title: Random Survival Forest
 doi: 10.5281/zenodo.5146376
-version: v0.1.1-beta
+version: v0.1.2-beta
