--- random_survival_forest/RandomSurvivalForest.py
@@ -1,151 +1,0 @@
-import multiprocessing
-
-import numpy as np
-import pandas as pd
-from joblib import Parallel, delayed
-
-from .SurvivalTree import SurvivalTree
-from .scoring import concordance_index
-
-
-class RandomSurvivalForest:
-
-    def __init__(self, n_estimators=100, min_leaf=3, unique_deaths=3,
-                 n_jobs=None, parallelization_backend="multiprocessing", oob_score=False):
-        """
-        A Random Survival Forest is a prediction model especially designed for survival analysis.
-        :param n_estimators: The numbers of trees in the forest.
-        :param min_leaf: The minimum number of samples required to be at a leaf node. A split point at any depth will
-        only be considered if it leaves at least min_leaf training samples in each of the left and right branches.
-        :param unique_deaths: The minimum number of unique deaths required to be at a leaf node.
-        :param n_jobs: The number of jobs to run in parallel for fit. None means 1.
-        """
-        self.n_estimators = n_estimators
-        self.min_leaf = min_leaf
-        self.unique_deaths = unique_deaths
-        self.n_jobs = n_jobs
-        self.parallelization_backend = parallelization_backend
-        self.bootstrap_idxs = None
-        self.bootstraps = []
-        self.oob_idxs = None
-        self.oob_score = oob_score
-        self.trees = []
-        self.timeline = None
-
-    def fit(self, x, y):
-        """
-        Build a forest of trees from the training set (X, y).
-        :param x: The input samples. Should be a Dataframe with the shape [n_samples, n_features].
-        :param y: The target values as a Dataframe with the survival time in the first column and the event
-        in the second with the shape [n_samples, 2]
-        :return: self: object
-        """
-        self.timeline = range(y.iloc[:, 1].min(), y.iloc[:, 1].max(), 1)
-        if self.n_jobs == -1:
-            self.n_jobs = multiprocessing.cpu_count()
-        elif self.n_jobs is None:
-            self.n_jobs = 1
-        self.bootstrap_idxs = self.draw_bootstrap_samples(x)
-
-        trees = Parallel(n_jobs=self.n_jobs, backend=self.parallelization_backend)(delayed(self.create_tree)(x, y, i)
-                                                                                   for i in range(self.n_estimators))
-
-        for i in range(len(trees)):
-            if trees[i].prediction_possible:
-                self.trees.append(trees[i])
-                self.bootstraps.append(self.bootstrap_idxs[i])
-
-        if self.oob_score:
-            self.oob_score = self.compute_oob_score(x, y)
-
-        return self
-
-    def create_tree(self, x, y, i):
-        """
-        Grows a survival tree for the bootstrap samples.
-        :param y: label data frame y with survival time as the first column and event as second
-        :param x: feature data frame x
-        :param i: Indices
-        :return: SurvivalTree
-        """
-        n_features = int(round(np.sqrt(x.shape[1]), 0))
-        f_idxs = np.random.permutation(x.shape[1])[:n_features]
-
-        tree = SurvivalTree(x=x.iloc[self.bootstrap_idxs[i], :], y=y.iloc[self.bootstrap_idxs[i], :],
-                            f_idxs=f_idxs, n_features=n_features,
-                            unique_deaths=self.unique_deaths, min_leaf=self.min_leaf,
-                            timeline=self.timeline)
-
-        return tree
-
-    def compute_oob_ensembles(self, xs):
-        """
-        Compute OOB ensembles.
-        :return: List of oob ensemble for each sample.
-        """
-        results = [compute_oob_ensemble_chf(sample_idx=sample_idx, xs=xs, trees=self.trees,
-                                            bootstraps=self.bootstraps) for sample_idx in range(xs.shape[0])]
-        oob_ensemble_chfs = [i for i in results if not i.empty]
-        return oob_ensemble_chfs
-
-    def compute_oob_score(self, x, y):
-        """
-        Compute the oob score (concordance-index).
-        :return: c-index of oob samples
-        """
-        oob_ensembles = self.compute_oob_ensembles(x)
-        c = concordance_index(y_time=y.iloc[:, 1], y_pred=oob_ensembles, y_event=y.iloc[:, 0])
-        return c
-
-    def predict(self, xs):
-        """
-        Predict survival for xs.
-        :param xs: The input samples
-        :return: List of the predicted cumulative hazard functions.
-        """
-        ensemble_chfs = [compute_ensemble_chf(sample_idx=sample_idx, xs=xs, trees=self.trees)
-                         for sample_idx in range(xs.shape[0])]
-        return ensemble_chfs
-
-    def draw_bootstrap_samples(self, data):
-        """
-        Draw bootstrap samples
-        :param data: Data to draw bootstrap samples of.
-        :return: Bootstrap indices for each of the trees
-        """
-        bootstrap_idxs = []
-        for i in range(self.n_estimators):
-            no_samples = len(data)
-            data_rows = range(no_samples)
-            bootstrap_idx = np.random.choice(data_rows, no_samples)
-            bootstrap_idxs.append(bootstrap_idx)
-
-        return bootstrap_idxs
-
-
-def compute_ensemble_chf(sample_idx, xs, trees):
-    denominator = 0
-    numerator = 0
-    for b in range(len(trees)):
-        sample = xs.iloc[sample_idx].to_list()
-        chf = trees[b].predict(sample)
-        denominator = denominator + 1
-        numerator = numerator + 1 * chf
-    ensemble_chf = numerator / denominator
-    return ensemble_chf
-
-
-def compute_oob_ensemble_chf(sample_idx, xs, trees, bootstraps):
-    denominator = 0
-    numerator = 0
-    for b in range(len(trees)):
-        if sample_idx not in bootstraps[b]:
-            sample = xs.iloc[sample_idx].to_list()
-            chf = trees[b].predict(sample)
-            denominator = denominator + 1
-            numerator = numerator + 1 * chf
-    if denominator != 0:
-        oob_ensemble_chf = numerator / denominator
-    else:
-        oob_ensemble_chf = pd.Series()
-    return oob_ensemble_chf
