+++ random_survival_forest/__init__.py
@@ -7,6 +7,4 @@ The annals of applied statistics, 2(3), 841-860.
 
 """
 
-from random_survival_forest.RandomSurvivalForest import RandomSurvivalForest
-from random_survival_forest.scoring import concordance_index
-from random_survival_forest.SurvivalTree import SurvivalTree
+from random_survival_forest import models, scoring
