+++ setup.cfg
@@ -1,3 +1,8 @@
-# Inside of setup.cfg
 [metadata]
 description-file = README.md
+
+[flake8]
+max-line-length = 120
+
+[pycodestyle]
+max-line-length = 120
